extract functions to avoid duplicate code
