+++ lib/Toolkit.py
@@ -8,6 +8,10 @@
 # Copyright (c) 2014-2015 	Pieter-Jan Moreels - pieterjan.moreels@gmail.com
 
 # Imports
+from dateutil import tz
+import dateutil.parser
+import re
+
 # Note of warning: CPEs like cpe:/o:microsoft:windows_8:-:-:x64 are given to us by Mitre
 #  x64 will be parsed as Edition in this case, not Architecture
 def toStringFormattedCPE(cpe,autofill=False):
@@ -79,3 +83,41 @@ def pad(seq, target_length, padding=None):
       return seq
     seq.extend([padding] * (target_length - length))
     return seq
+
+def currentTime(utc):
+    timezone = tz.tzlocal()
+    utc = dateutil.parser.parse(utc)
+    output = utc.astimezone(timezone)
+    output = output.strftime('%d-%m-%Y - %H:%M')
+    return output
+
+def isURL(string):
+    urlTypes= [re.escape(x) for x in ['http://','https://', 'www.']]
+    return re.match("^(" + "|".join(urlTypes) + ")", string)
+
+def vFeedName(string):
+    string=string.replace('map_','')
+    string=string.replace('cve_','')
+    return string.title()
+
+def convertDateToDBFormat(string):
+    result = None
+    try:
+        result = time.strptime(string, "%d-%m-%Y")
+    except:
+        pass
+    try:
+        result = time.strptime(string, "%d-%m-%y")
+    except:
+        pass
+    try:
+        result = time.strptime(string, "%d/%m/%Y")
+    except:
+        pass
+    try:
+        result = time.strptime(string, "%d/%m/%y")
+    except:
+        pass
+    if result is not None:
+        result = time.strftime('%Y-%m-%d', result)
+    return result
