+++ web/index.py
@@ -27,8 +27,6 @@ from passlib.hash import pbkdf2_sha256
 from redis import exceptions as redisExceptions
 
 import json
-from dateutil import tz
-import dateutil.parser
 import re
 import argparse
 import time
@@ -41,7 +39,7 @@ from logging.handlers import RotatingFileHandler
 
 from lib.User import User
 from lib.Config import Configuration
-from lib.Toolkit import toStringFormattedCPE, toOldCPE
+from lib.Toolkit import toStringFormattedCPE, toOldCPE, currentTime, isURL, vFeedName, convertDateToDBFormat
 import lib.CVEs as cves
 from bin.db_whitelist import *
 from bin.db_blacklist import *
@@ -160,9 +158,6 @@ def getBlacklistRegexes():
 
 
 def addCPEToList(cpe, listType, cpeType=None):
-    #cpe = urllib.parse.quote_plus(cpe).lower()
-    #cpe = cpe.replace("%3a", ":")
-    #cpe = cpe.replace("%2f", "/")
     if not cpeType:
         cpeType='cpe'
     if listType.lower() in ("blacklist", "black", "b", "bl"):
@@ -180,29 +175,6 @@ def addCPEToList(cpe, listType, cpeType=None):
 def getVersionsOfProduct(product):
     p = redisdb.smembers("p:" + product)
     return sorted(list(p))
-
-
-def convertDateToDBFormat(string):
-    result = None
-    try:
-        result = time.strptime(string, "%d-%m-%Y")
-    except:
-        pass
-    try:
-        result = time.strptime(string, "%d-%m-%y")
-    except:
-        pass
-    try:
-        result = time.strptime(string, "%d/%m/%Y")
-    except:
-        pass
-    try:
-        result = time.strptime(string, "%d/%m/%y")
-    except:
-        pass
-    if result is not None:
-        result = time.strftime('%Y-%m-%d', result)
-    return result
 
 
 def adminStats():
@@ -804,12 +776,8 @@ def page_not_found(e):
 
 # filters
 @app.template_filter('currentTime')
-def currentTime(utc):
-    timezone = tz.tzlocal()
-    utc = dateutil.parser.parse(utc)
-    output = utc.astimezone(timezone)
-    output = output.strftime('%d-%m-%Y - %H:%M')
-    return output
+def currentTimeFilter(utc):
+    return currentTime(utc)
 
 
 @app.template_filter('htmlDecode')
@@ -822,25 +790,13 @@ def htmlEncode(string):
     return urllib.parse.quote_plus(string).lower()
 
 
-@app.template_filter('impact')
-def impact(string):
-    if string.lower() == "none":
-        return "good"
-    elif string.lower() == "partial":
-        return "medium"
-    elif string.lower() == "complete":
-        return "bad"
-
 @app.template_filter('isURL')
-def isURL(string):
-    urlTypes= [re.escape(x) for x in ['http://','https://', 'www.']]
-    return re.match("^(" + "|".join(urlTypes) + ")", string)
+def isURLFilter(string):
+    return isURL(string)
 
 @app.template_filter('vFeedName')
-def vFeedName(string):
-    string=string.replace('map_','')
-    string=string.replace('cve_','')
-    return string.title()
+def vFeedNameFilter(string):
+    return vFeedName(string)
 
 # signal handlers
 def sig_handler(sig, frame):
