+++ web/minimal-web.py
@@ -24,8 +24,6 @@ from flask.ext.pymongo import PyMongo
 from redis import exceptions as redisExceptions
 
 import json
-from dateutil import tz
-import dateutil.parser
 import re
 import argparse
 import time
@@ -36,7 +34,7 @@ import logging
 from logging.handlers import RotatingFileHandler
 
 from lib.Config import Configuration
-from lib.Toolkit import toStringFormattedCPE, toOldCPE
+from lib.Toolkit import toStringFormattedCPE, toOldCPE, currentTime, isURL, vFeedName, convertDateToDBFormat
 import lib.CVEs as cves
 
 # parse command line arguments
@@ -75,33 +73,9 @@ def getBrowseList(vendor):
     return result
 
 
-
 def getVersionsOfProduct(product):
     p = redisdb.smembers("p:" + product)
     return sorted(list(p))
-
-
-def convertDateToDBFormat(string):
-    result = None
-    try:
-        result = time.strptime(string, "%d-%m-%Y")
-    except:
-        pass
-    try:
-        result = time.strptime(string, "%d-%m-%y")
-    except:
-        pass
-    try:
-        result = time.strptime(string, "%d/%m/%Y")
-    except:
-        pass
-    try:
-        result = time.strptime(string, "%d/%m/%y")
-    except:
-        pass
-    if result is not None:
-        result = time.strftime('%Y-%m-%d', result)
-    return result
 
 
     cveU = db.info.find_one({'db': 'cve'})
@@ -220,7 +194,6 @@ def filterLast(r):
     # retrieving data
     cve = filter_logic(unlisted, timeSelect, startDate, endDate,
                        timeTypeSelect, cvssSelect, cvss, rejectedSelect, pageLength, r)
-
     return render_template('index-minimal.html', settings=settings, cve=cve, r=r, pageLength=pageLength)
 
 @app.route('/api/cpe2.3/<path:cpe>', methods=['GET'])
@@ -335,12 +308,8 @@ def page_not_found(e):
 
 # filters
 @app.template_filter('currentTime')
-def currentTime(utc):
-    timezone = tz.tzlocal()
-    utc = dateutil.parser.parse(utc)
-    output = utc.astimezone(timezone)
-    output = output.strftime('%d-%m-%Y - %H:%M')
-    return output
+def currentTimeFilter(utc):
+    return currentTime(utc)
 
 
 @app.template_filter('htmlDecode')
@@ -353,25 +322,13 @@ def htmlEncode(string):
     return urllib.parse.quote_plus(string).lower()
 
 
-@app.template_filter('impact')
-def impact(string):
-    if string.lower() == "none":
-        return "good"
-    elif string.lower() == "partial":
-        return "medium"
-    elif string.lower() == "complete":
-        return "bad"
-
 @app.template_filter('isURL')
-def isURL(string):
-    urlTypes= [re.escape(x) for x in ['http://','https://', 'www.']]
-    return re.match("^(" + "|".join(urlTypes) + ")", string)
+def isURLFilter(string):
+    return isURL(string)
 
 @app.template_filter('vFeedName')
-def vFeedName(string):
-    string=string.replace('map_','')
-    string=string.replace('cve_','')
-    return string.title()
+def vFeedNameFilter(string):
+    return vFeedName(string)
 
 # signal handlers
 def sig_handler(sig, frame):
