+++ web/static/css/style.css
@@ -71,14 +71,14 @@ ul.nav li.dropdown:hover > ul.dropdown-menu {
    table-layout: fixed
 }
 
-.good {
+.impact-none {
    color:green;
 }
 
-.medium {
+.impact-partial {
    color:orange;
 }
 
-.bad {
+.impact-complete {
    color:red;
 }
