+++ web/templates/cve.html
@@ -136,9 +136,9 @@
                     <table class="table table-hover table-bordered cve-info">
                       <thead><tr class="warning"><td>Confidentiality</td><td>Integrity</td><td>Availability</td></tr></thead>
                       <tr>
-                        <td class="{{cve['impact']['confidentiality']|impact}}">{{cve['impact']['confidentiality']}}</td>
-                        <td class="{{cve['impact']['integrity']|impact}}">{{cve['impact']['integrity']}}</td>
-                        <td class="{{cve['impact']['availability']|impact}}">{{cve['impact']['availability']}}</td>
+                        <td class="impact-{{cve['impact']['confidentiality']|lower}}">{{cve['impact']['confidentiality']}}</td>
+                        <td class="impact-{{cve['impact']['integrity']|lower}}">{{cve['impact']['integrity']}}</td>
+                        <td class="impact-{{cve['impact']['availability']|lower}}">{{cve['impact']['availability']}}</td>
                       </tr>
                     </table>
                   </td>
